Reorganize files + rename to ChatChannelUserDetailVC
