+++ Sources_v3/StreamChatUI/ChatChannelDetail/ChatChannelUserDetailCollectionView.swift
@@ -1,0 +1,15 @@
+//
+// Copyright © 2021 Stream.io Inc. All rights reserved.
+//
+
+import UIKit
+
+open class ChatChannelUserDetailCollectionView: UICollectionView {
+    public required init(layout: UICollectionViewLayout) {
+        super.init(frame: .zero, collectionViewLayout: layout)
+    }
+
+    public required init?(coder: NSCoder) {
+        super.init(coder: coder)
+    }
+}
